Decouple ObjectManager from Map
