+++ GameServerLib/Logic/Chatbox/Commands/ChCommand.cs
@@ -38,8 +38,8 @@ namespace LeagueSandbox.GameServer.Logic.Chatbox.Commands
             );
             c.Model = split[1]; // trigger the "modelUpdate" proc
             c.SetTeam(_playerManager.GetPeerInfo(peer).Champion.Team);
-            _game.Map.RemoveObject(_playerManager.GetPeerInfo(peer).Champion);
-            _game.Map.AddObject(c);
+            _game.ObjectManager.RemoveObject(_playerManager.GetPeerInfo(peer).Champion);
+            _game.ObjectManager.AddObject(c);
             _playerManager.GetPeerInfo(peer).Champion = c;
         }
     }
