+++ GameServerLib/Logic/GameObjects/AttackableUnits/Buildings/AnimatedBuildings/Inhibitor.cs
@@ -32,7 +32,7 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
 
         public override void die(Unit killer)
         {
-            var objects = _game.Map.GetObjects().Values;
+            var objects = _game.ObjectManager.GetObjects().Values;
             foreach (var obj in objects)
             {
                 var u = obj as Unit;
